requirements.txt
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-image==0.21.0
numpy==1.24.3